implementação dos filtros e da ordenação
