+++ src/components/estiloBuscaOrdenacao.js
@@ -1,0 +1,7 @@
+import styled from "styled-components"
+
+
+export const BuscaOrdenacaoContainer = styled.div`
+    display:flex;
+    justify-content: space-between;
+`
