+++ src/components/estiloDoFiltro.js
@@ -1,12 +1,17 @@
 import styled from 'styled-components'
 export const GrupoDeFiltros = styled.div` 
-    border: 1px solid black;
     display: flex;
-    flex-direction: column;
+    justify-content: center;
     padding: 10px;
 `
 export const Filtro = styled.label`
     display: flex;
     flex-direction: column;
+    align-items: center;
     padding: 10px 0px;
+    color:white;
+    
+`
+export const BoxFiltro = styled.input`
+    height: 5%;
 `
