+++ src/components/estiloOrdenacao.js
@@ -1,0 +1,12 @@
+import styled from 'styled-components'
+
+export const EstiloOrdenacao = styled.select`
+    height: 8%;
+    align-self: center;
+    
+`
+export const OrdenacaoContainer = styled.div`
+    color:white;
+    align-self:center;
+    align-items: center;
+`
